% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLHD.R
\name{rLHD}
\alias{rLHD}
\title{Generate a random Latin Hypercube Design (LHD)}
\usage{
rLHD(n, k)
}
\arguments{
\item{n}{A positive integer.}

\item{k}{A positive integer.}
}
\value{
If all inputs are positive integer, then the output will be a \code{n} by \code{k} design matrix.
}
\description{
\code{rLHD} returns an Latin hypercube design matrix with user-defined dimention
}
\details{
\itemize{
\item \code{n} stands for the number of rows (or run size).
\item \code{k} stands for the number of columns (or the number of factors).
}
}
\examples{
#create a toy LHD with 5 rows and 3 columns
toy=rLHD(n=5,k=3);toy

#another example with 9 rows and 2 columns
rLHD(9,2)
}
